Update Eldrow Game.py
Implements the following:
- The main function, which serves to run the game & contains all game-logic.
- Formatting for the screen & time-delays.

Developer Comment: Comprising all game-logic, the function looks at four cases. Firstly, it accepts a 'letter-guess' -- which is checked against a set to ensure a letter is not guessed twice. If it is not, the letter is added to the guess. From there, it'll check the occurrences of the letter within the secret-word and update accordingly. Next, it accepts a 'correct-word' case, where as the name implies, the guess is correct. Next, it accepts the first 'wrong-word' case, which describes if the length of the guess is not equal to the length of the chosen word, and is not the chosen word. Finally, it accepts a secondary 'wrong-word' case, which is like the first one, just if the length of the guess equals the chosen word (i.e white and black, banana and orange, etc.) At the end, it reveals the word, if not found, and number of letter guesses.

Note: Pasted from VSCode after development & extensive testing to ensure program functions before commit.